# api_server/openapi_server/models/build_status.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class BuildStatus(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, branch=None, git_source=None, organization=None, project_id=None, repo_name=None, status=None, updated=None):  # noqa: E501
        """BuildStatus - a model defined in OpenAPI

        :param branch: The branch of this BuildStatus.  # noqa: E501
        :type branch: str
        :param git_source: The git_source of this BuildStatus.  # noqa: E501
        :type git_source: str
        :param organization: The organization of this BuildStatus.  # noqa: E501
        :type organization: str
        :param project_id: The project_id of this BuildStatus.  # noqa: E501
        :type project_id: str
        :param repo_name: The repo_name of this BuildStatus.  # noqa: E501
        :type repo_name: str
        :param status: The status of this BuildStatus.  # noqa: E501
        :type status: str
        :param updated: The updated of this BuildStatus.  # noqa: E501
        :type updated: datetime
        """
        self.openapi_types = {
            'branch': str,
            'git_source': str,
            'organization': str,
            'project_id': str,
            'repo_name': str,
            'status': str,
            'updated': datetime
        }

        self.attribute_map = {
            'branch': 'branch',
            'git_source': 'git_source',
            'organization': 'organization',
            'project_id': 'project_id',
            'repo_name': 'repo_name',
            'status': 'status',
            'updated': 'updated'
        }

        self._branch = branch
        self._git_source = git_source
        self._organization = organization
        self._project_id = project_id
        self._repo_name = repo_name
        self._status = status
        self._updated = updated

    @classmethod
    def from_dict(cls, dikt) -> 'BuildStatus':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The BuildStatus of this BuildStatus.  # noqa: E501
        :rtype: BuildStatus
        """
        return util.deserialize_model(dikt, cls)

    @property
    def branch(self):
        """Gets the branch of this BuildStatus.


        :return: The branch of this BuildStatus.
        :rtype: str
        """
        return self._branch

    @branch.setter
    def branch(self, branch):
        """Sets the branch of this BuildStatus.


        :param branch: The branch of this BuildStatus.
        :type branch: str
        """
        if branch is None:
            raise ValueError("Invalid value for `branch`, must not be `None`")  # noqa: E501

        self._branch = branch

    @property
    def git_source(self):
        """Gets the git_source of this BuildStatus.


        :return: The git_source of this BuildStatus.
        :rtype: str
        """
        return self._git_source

    @git_source.setter
    def git_source(self, git_source):
        """Sets the git_source of this BuildStatus.


        :param git_source: The git_source of this BuildStatus.
        :type git_source: str
        """
        if git_source is None:
            raise ValueError("Invalid value for `git_source`, must not be `None`")  # noqa: E501

        self._git_source = git_source

    @property
    def organization(self):
        """Gets the organization of this BuildStatus.


        :return: The organization of this BuildStatus.
        :rtype: str
        """
        return self._organization

    @organization.setter
    def organization(self, organization):
        """Sets the organization of this BuildStatus.


        :param organization: The organization of this BuildStatus.
        :type organization: str
        """
        if organization is None:
            raise ValueError("Invalid value for `organization`, must not be `None`")  # noqa: E501

        self._organization = organization

    @property
    def project_id(self):
        """Gets the project_id of this BuildStatus.


        :return: The project_id of this BuildStatus.
        :rtype: str
        """
        return self._project_id

    @project_id.setter
    def project_id(self, project_id):
        """Sets the project_id of this BuildStatus.


        :param project_id: The project_id of this BuildStatus.
        :type project_id: str
        """
        if project_id is None:
            raise ValueError("Invalid value for `project_id`, must not be `None`")  # noqa: E501

        self._project_id = project_id

    @property
    def repo_name(self):
        """Gets the repo_name of this BuildStatus.


        :return: The repo_name of this BuildStatus.
        :rtype: str
        """
        return self._repo_name

    @repo_name.setter
    def repo_name(self, repo_name):
        """Sets the repo_name of this BuildStatus.


        :param repo_name: The repo_name of this BuildStatus.
        :type repo_name: str
        """
        if repo_name is None:
            raise ValueError("Invalid value for `repo_name`, must not be `None`")  # noqa: E501

        self._repo_name = repo_name

    @property
    def status(self):
        """Gets the status of this BuildStatus.


        :return: The status of this BuildStatus.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status):
        """Sets the status of this BuildStatus.


        :param status: The status of this BuildStatus.
        :type status: str
        """
        allowed_values = ["pending", "failing", "passing"]  # noqa: E501
        if status not in allowed_values:
            raise ValueError(
                "Invalid value for `status` ({0}), must be one of {1}"
                .format(status, allowed_values)
            )

        self._status = status

    @property
    def updated(self):
        """Gets the updated of this BuildStatus.


        :return: The updated of this BuildStatus.
        :rtype: datetime
        """
        return self._updated

    @updated.setter
    def updated(self, updated):
        """Sets the updated of this BuildStatus.


        :param updated: The updated of this BuildStatus.
        :type updated: datetime
        """
        if updated is None:
            raise ValueError("Invalid value for `updated`, must not be `None`")  # noqa: E501

        self._updated = updated